refactor: split systems into separate modules
