+++ rustyhack_server/src/game/systems/common_entity_systems.rs
@@ -1,0 +1,50 @@
+use legion::world::SubWorld;
+use legion::{system, Query};
+use rustyhack_lib::background_map::tiles::{Collidable, Tile};
+use rustyhack_lib::background_map::{AllMaps, BackgroundMap};
+use rustyhack_lib::consts::DEFAULT_MAP;
+use rustyhack_lib::ecs::components::Position;
+use rustyhack_lib::math_utils::{i32_from, u32_from};
+
+#[system]
+pub(crate) fn update_entities_position(
+    world: &mut SubWorld,
+    query: &mut Query<&mut Position>,
+    #[resource] all_maps: &AllMaps,
+) {
+    for position in query.iter_mut(world) {
+        debug!("Checking for possible movement after velocity updates and combat check.");
+        if position.velocity_x == 0 && position.velocity_y == 0 {
+            //no velocity, no updates
+            continue;
+        }
+        let current_map = get_current_map(all_maps, &position.current_map);
+        let potential_pos_x = u32_from(i32_from(position.pos_x) + position.velocity_x);
+        let potential_pos_y = u32_from(i32_from(position.pos_y) + position.velocity_y);
+
+        if !entity_is_colliding_with_tile(current_map.get_tile_at(potential_pos_x, potential_pos_y))
+        {
+            position.pos_x = potential_pos_x;
+            position.pos_y = potential_pos_y;
+        }
+        position.velocity_x = 0;
+        position.velocity_y = 0;
+    }
+}
+
+fn get_current_map<'a>(all_maps: &'a AllMaps, map: &String) -> &'a BackgroundMap {
+    all_maps.get(map).unwrap_or_else(|| {
+        error!("Entity is located on a map that does not exist: {}", &map);
+        warn!("Will return the default map, but this may cause problems.");
+        all_maps.get(DEFAULT_MAP).unwrap()
+    })
+}
+
+fn entity_is_colliding_with_tile(tile: Tile) -> bool {
+    match tile {
+        Tile::Door(door) => door.collidable == Collidable::True,
+        Tile::Wall(wall) => wall.collidable == Collidable::True,
+        Tile::Boundary => true,
+        _ => false,
+    }
+}
